bots/officebot/requirements.txt
python-telegram-bot[job-queue]>=20.0
httpx>=0.25.0
aiohttp>=3.9.0
